Throw an exception instead of passing as a callback parameter
DotNet runtime fills a stack trace in an Exception object when it is thrown. Unless throwing, it's kept as empty.
Some libraries which refer a stack trace in an Exception causes error when a stack trace is empty.
To avoid causing such errors, an API exception is thrown and caught to fill a stack trace instead of passing as a callback parameter directly.